pyproject.toml
[project]
name = "ai-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]
